Recompile nav.css using the latest theme templates
Now that we've moved the light, dark, and green themes to the new theme
template, it's time to recompile. Very little has changed in the compiled
output. The dark theme uses line-height instead of padding. The light
theme uses slightly different (LESS-generated) colors. Some of the
selecters were normalized, and a few of the properties were reorganized.