Switch to the latest Flexbox (Fixes #21)
This update adds support for the newer flexbox syntax, supported by
Chrome, IE10, and Opera 12.10. Unfortunately, Safari and Firefox do not
yet support the latest version of the specification. Firefox will do so in March... not sure about Safari.

Modernizr was also updated to a newer version that correctly tests flexbox
on Chrome (and no longer tests flexbox-legacy).