Fix mega menus with border-radius
This indents mega menus so that they start/end where the roundedness
begins.  Unfortunately we have to set a fixed width of 100% for mega
menus in IE9 because we're using display: table there, which means that
setting a left value would indent the entire menu too far. So we're
doing this for only flexbox-friendly browsers.