Make navbar-cover transition long enough for touch
It seems like the only way around the touch persistence problem is
having a long enough delay on the cover element so the touch stays away
from any links on the page. It's dirty but that's what it looks like we
have to do. 350ms seems to work well and it's not too slow. Also slowed
down the left transition to make it more equal.